Fix parameters of FloatInRange
Introduced in f6cd786d1.

Reported-by: Jean-Pierre Demailly <Jean-Pierre.Demailly@ujf-grenoble.fr>

+++ src/Scrollbar.c
@@ -865,7 +865,7 @@ InRange(int num, int small, int big)
  */
 
 static float
-FloatInRange(int num, int small, int big)
+FloatInRange(float num, float small, float big)
 {
     return (num < small) ? small : ((num > big) ? big : num);
 }
